@@ -12,7 +12,6 @@ import (
 	reflect "reflect"
 	strings "strings"
 	v11 "github.com/temporalio/tcld/protogen/api/auth/v1"
-	v14 "github.com/temporalio/tcld/protogen/api/common/v1"
 	v1 "github.com/temporalio/tcld/protogen/api/namespace/v1"
 	v12 "github.com/temporalio/tcld/protogen/api/request/v1"
 	v13 "github.com/temporalio/tcld/protogen/api/sink/v1"
@@ -1456,28 +1455,24 @@ func (m *GetExportSinksResponse) GetNextPageToken() string {
 	return ""
 }
 
-type FailoverNamespaceRequest struct {
-	// The namespace to be failed over.
+type ValidateExportSinkRequest struct {
+	// the namespace the sink is scoped to
 	Namespace string `protobuf:"bytes,1,opt,name=namespace,proto3" json:"namespace,omitempty"`
-	// the request id to use for this operation - optional
-	RequestId string `protobuf:"bytes,2,opt,name=request_id,json=requestId,proto3" json:"request_id,omitempty"`
-	// The target region is the region to be primary after the failover.
-	TargetRegion *v14.RegionID `protobuf:"bytes,3,opt,name=target_region,json=targetRegion,proto3" json:"target_region,omitempty"`
-	// This will skip graceful failover and fail over namespace without any condition.
-	SkipGracefulFailover bool `protobuf:"varint,4,opt,name=skip_graceful_failover,json=skipGracefulFailover,proto3" json:"skip_graceful_failover,omitempty"`
+	// the sink specification needs to be validated
+	Spec *v13.ExportSinkSpec `protobuf:"bytes,2,opt,name=spec,proto3" json:"spec,omitempty"`
 }
 
-func (m *FailoverNamespaceRequest) Reset()      { *m = FailoverNamespaceRequest{} }
-func (*FailoverNamespaceRequest) ProtoMessage() {}
-func (*FailoverNamespaceRequest) Descriptor() ([]byte, []int) {
+func (m *ValidateExportSinkRequest) Reset()      { *m = ValidateExportSinkRequest{} }
+func (*ValidateExportSinkRequest) ProtoMessage() {}
+func (*ValidateExportSinkRequest) Descriptor() ([]byte, []int) {
 	return fileDescriptor_667e39c23eb47b7c, []int{26}
 }
-func (m *FailoverNamespaceRequest) XXX_Unmarshal(b []byte) error {
+func (m *ValidateExportSinkRequest) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
 }
-func (m *FailoverNamespaceRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+func (m *ValidateExportSinkRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
 	if deterministic {
-		return xxx_messageInfo_FailoverNamespaceRequest.Marshal(b, m, deterministic)
+		return xxx_messageInfo_ValidateExportSinkRequest.Marshal(b, m, deterministic)
 	} else {
 		b = b[:cap(b)]
 		n, err := m.MarshalToSizedBuffer(b)
@@ -1487,62 +1482,46 @@ func (m *FailoverNamespaceRequest) XXX_Marshal(b []byte, deterministic bool) ([]
 		return b[:n], nil
 	}
 }
-func (m *FailoverNamespaceRequest) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_FailoverNamespaceRequest.Merge(m, src)
+func (m *ValidateExportSinkRequest) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_ValidateExportSinkRequest.Merge(m, src)
 }
-func (m *FailoverNamespaceRequest) XXX_Size() int {
+func (m *ValidateExportSinkRequest) XXX_Size() int {
 	return m.Size()
 }
-func (m *FailoverNamespaceRequest) XXX_DiscardUnknown() {
-	xxx_messageInfo_FailoverNamespaceRequest.DiscardUnknown(m)
+func (m *ValidateExportSinkRequest) XXX_DiscardUnknown() {
+	xxx_messageInfo_ValidateExportSinkRequest.DiscardUnknown(m)
 }
 
-var xxx_messageInfo_FailoverNamespaceRequest proto.InternalMessageInfo
+var xxx_messageInfo_ValidateExportSinkRequest proto.InternalMessageInfo
 
-func (m *FailoverNamespaceRequest) GetNamespace() string {
+func (m *ValidateExportSinkRequest) GetNamespace() string {
 	if m != nil {
 		return m.Namespace
 	}
 	return ""
 }
 
-func (m *FailoverNamespaceRequest) GetRequestId() string {
+func (m *ValidateExportSinkRequest) GetSpec() *v13.ExportSinkSpec {
 	if m != nil {
-		return m.RequestId
-	}
-	return ""
-}
-
-func (m *FailoverNamespaceRequest) GetTargetRegion() *v14.RegionID {
-	if m != nil {
-		return m.TargetRegion
+		return m.Spec
 	}
 	return nil
 }
 
-func (m *FailoverNamespaceRequest) GetSkipGracefulFailover() bool {
-	if m != nil {
-		return m.SkipGracefulFailover
-	}
-	return false
+type ValidateExportSinkResponse struct {
 }
 
-type FailoverNamespaceResponse struct {
-	// Failover status
-	RequestStatus *v12.RequestStatus `protobuf:"bytes,1,opt,name=request_status,json=requestStatus,proto3" json:"request_status,omitempty"`
-}
-
-func (m *FailoverNamespaceResponse) Reset()      { *m = FailoverNamespaceResponse{} }
-func (*FailoverNamespaceResponse) ProtoMessage() {}
-func (*FailoverNamespaceResponse) Descriptor() ([]byte, []int) {
+func (m *ValidateExportSinkResponse) Reset()      { *m = ValidateExportSinkResponse{} }
+func (*ValidateExportSinkResponse) ProtoMessage() {}
+func (*ValidateExportSinkResponse) Descriptor() ([]byte, []int) {
 	return fileDescriptor_667e39c23eb47b7c, []int{27}
 }
-func (m *FailoverNamespaceResponse) XXX_Unmarshal(b []byte) error {
+func (m *ValidateExportSinkResponse) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
 }
-func (m *FailoverNamespaceResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+func (m *ValidateExportSinkResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
 	if deterministic {
-		return xxx_messageInfo_FailoverNamespaceResponse.Marshal(b, m, deterministic)
+		return xxx_messageInfo_ValidateExportSinkResponse.Marshal(b, m, deterministic)
 	} else {
 		b = b[:cap(b)]
 		n, err := m.MarshalToSizedBuffer(b)
@@ -1552,24 +1531,18 @@ func (m *FailoverNamespaceResponse) XXX_Marshal(b []byte, deterministic bool) ([
 		return b[:n], nil
 	}
 }
-func (m *FailoverNamespaceResponse) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_FailoverNamespaceResponse.Merge(m, src)
+func (m *ValidateExportSinkResponse) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_ValidateExportSinkResponse.Merge(m, src)
 }
-func (m *FailoverNamespaceResponse) XXX_Size() int {
+func (m *ValidateExportSinkResponse) XXX_Size() int {
 	return m.Size()
 }
-func (m *FailoverNamespaceResponse) XXX_DiscardUnknown() {
-	xxx_messageInfo_FailoverNamespaceResponse.DiscardUnknown(m)
+func (m *ValidateExportSinkResponse) XXX_DiscardUnknown() {
+	xxx_messageInfo_ValidateExportSinkResponse.DiscardUnknown(m)
 }
 
-var xxx_messageInfo_FailoverNamespaceResponse proto.InternalMessageInfo
+var xxx_messageInfo_ValidateExportSinkResponse proto.InternalMessageInfo
 
-func (m *FailoverNamespaceResponse) GetRequestStatus() *v12.RequestStatus {
-	if m != nil {
-		return m.RequestStatus
-	}
-	return nil
-}
 
 func init() {
 	proto.RegisterType((*CreateNamespaceRequest)(nil), "api.namespaceservice.v1.CreateNamespaceRequest")
@@ -1598,8 +1571,8 @@ func init() {
 	proto.RegisterType((*ListExportSinksResponse)(nil), "api.namespaceservice.v1.ListExportSinksResponse")
 	proto.RegisterType((*GetExportSinksRequest)(nil), "api.namespaceservice.v1.GetExportSinksRequest")
 	proto.RegisterType((*GetExportSinksResponse)(nil), "api.namespaceservice.v1.GetExportSinksResponse")
-	proto.RegisterType((*FailoverNamespaceRequest)(nil), "api.namespaceservice.v1.FailoverNamespaceRequest")
-	proto.RegisterType((*FailoverNamespaceResponse)(nil), "api.namespaceservice.v1.FailoverNamespaceResponse")
+	proto.RegisterType((*ValidateExportSinkRequest)(nil), "api.namespaceservice.v1.ValidateExportSinkRequest")
+	proto.RegisterType((*ValidateExportSinkResponse)(nil), "api.namespaceservice.v1.ValidateExportSinkResponse")
 }
 
 func init() {
@@ -1607,66 +1580,67 @@ func init() {
 }
 
 var fileDescriptor_667e39c23eb47b7c = []byte{
-	// 942 bytes of a gzipped FileDescriptorProto
+	// 958 bytes of a gzipped FileDescriptorProto
 	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xbc, 0x57, 0x4f, 0x6f, 0x1b, 0x45,
-	0x14, 0xf7, 0xd8, 0x0e, 0x8a, 0x5f, 0x08, 0x45, 0x4b, 0xeb, 0x6c, 0xf3, 0x67, 0x31, 0x8b, 0x8a,
-	0x82, 0x10, 0xb6, 0x92, 0xf4, 0x82, 0xca, 0x05, 0x1a, 0x5a, 0x55, 0x42, 0x28, 0xac, 0x29, 0x07,
-	0x2e, 0x9b, 0xed, 0xe6, 0xd5, 0x1d, 0x6c, 0xef, 0x6e, 0x67, 0x66, 0xdd, 0xa8, 0xe2, 0xd0, 0x1b,
-	0x57, 0x3e, 0x01, 0x47, 0xc4, 0x19, 0x89, 0xef, 0x80, 0xc4, 0x25, 0xc7, 0x1e, 0x89, 0x73, 0xe1,
-	0x46, 0x3f, 0x02, 0x9a, 0xd9, 0x1d, 0xff, 0xd9, 0xb5, 0xc3, 0x16, 0x6d, 0x7b, 0xf4, 0x7b, 0xcf,
-	0xbf, 0xf7, 0xfb, 0xbd, 0x79, 0x6f, 0xf6, 0x0d, 0xb4, 0xbd, 0x88, 0x76, 0x02, 0x6f, 0x88, 0x3c,
-	0xf2, 0x7c, 0xe4, 0xc8, 0x46, 0xd4, 0xc7, 0xce, 0x68, 0xaf, 0xc3, 0xf0, 0x71, 0x8c, 0x5c, 0xb8,
-	0x0c, 0x79, 0x14, 0x06, 0x1c, 0xdb, 0x11, 0x0b, 0x45, 0x68, 0x6c, 0x78, 0x11, 0x6d, 0x67, 0xe3,
-	0xdb, 0xa3, 0xbd, 0xcd, 0xeb, 0x12, 0xc8, 0x8b, 0xc5, 0x23, 0xf9, 0xe7, 0x21, 0x72, 0xee, 0xf5,
-	0xd2, 0xff, 0x6c, 0x6e, 0x49, 0x97, 0x1f, 0x0e, 0x87, 0x61, 0x90, 0x77, 0x5a, 0x73, 0x04, 0xf2,
-	0xfe, 0x6d, 0xe9, 0x4f, 0xc9, 0xe4, 0xbd, 0x2a, 0x2b, 0xa7, 0x41, 0x3f, 0xe7, 0xb2, 0xff, 0x21,
-	0xd0, 0xbc, 0xcd, 0xd0, 0x13, 0xf8, 0x95, 0x46, 0x77, 0x12, 0x18, 0x63, 0x1b, 0x1a, 0x93, 0x8c,
-	0x26, 0x69, 0x91, 0xdd, 0x86, 0x33, 0x35, 0x18, 0x07, 0x50, 0xe7, 0x11, 0xfa, 0x66, 0xb5, 0x45,
-	0x76, 0xd7, 0xf6, 0xdf, 0x6d, 0xcf, 0x29, 0x6e, 0x8f, 0xf6, 0xda, 0x13, 0xbc, 0x6e, 0x84, 0xbe,
-	0xa3, 0x82, 0x8d, 0x1d, 0x00, 0x5d, 0x31, 0x7a, 0x62, 0xd6, 0x12, 0xcc, 0xd4, 0x72, 0xef, 0xc4,
-	0xf0, 0x61, 0x33, 0xe6, 0xc8, 0xdc, 0x09, 0x8e, 0x1b, 0x21, 0x1b, 0x52, 0xce, 0x69, 0x18, 0x70,
-	0xb3, 0xde, 0xaa, 0xed, 0xae, 0xed, 0xdf, 0x50, 0x99, 0x64, 0x09, 0x65, 0x92, 0xfb, 0x1c, 0xd9,
-	0x24, 0xd1, 0xd1, 0x34, 0xd8, 0x31, 0xe3, 0x25, 0x1e, 0xdb, 0x85, 0x8d, 0x9c, 0xe0, 0xe4, 0xf0,
-	0x8c, 0x43, 0x78, 0x4b, 0xd3, 0xe3, 0xc2, 0x13, 0x31, 0x57, 0xb2, 0xd7, 0xf6, 0x77, 0x54, 0xce,
-	0xd4, 0x25, 0xd3, 0xa6, 0x25, 0xea, 0xaa, 0x20, 0x67, 0x9d, 0xcd, 0xfe, 0xb4, 0xbb, 0x70, 0xed,
-	0x4b, 0xca, 0xc5, 0x04, 0x9e, 0xeb, 0x82, 0x6e, 0x41, 0x23, 0xf2, 0x7a, 0xe8, 0x72, 0xfa, 0x34,
-	0x29, 0xe8, 0x8a, 0xb3, 0x2a, 0x0d, 0x5d, 0xfa, 0x14, 0x65, 0x69, 0x94, 0x53, 0x84, 0x7d, 0x0c,
-	0x54, 0x55, 0x1b, 0x8e, 0x0a, 0xff, 0x46, 0x1a, 0xec, 0x63, 0x68, 0x66, 0x41, 0x53, 0xd2, 0x16,
-	0xc0, 0xb4, 0xd3, 0x4c, 0xd2, 0xaa, 0xed, 0x36, 0x9c, 0x19, 0x8b, 0xf1, 0x01, 0x5c, 0x09, 0xf0,
-	0x54, 0xb8, 0x39, 0xf4, 0x75, 0x69, 0x3e, 0x9a, 0x64, 0x70, 0xe0, 0xea, 0x5d, 0x2c, 0x99, 0xf5,
-	0x0f, 0x70, 0x2d, 0x83, 0x99, 0x92, 0xbe, 0x95, 0x23, 0xbd, 0xb6, 0xbf, 0x75, 0x49, 0x0f, 0xfd,
-	0x2f, 0x45, 0x07, 0xf0, 0xce, 0x6c, 0xf6, 0x42, 0x7d, 0x6d, 0x7f, 0x3d, 0x5f, 0x86, 0x09, 0xe3,
-	0x4f, 0xb2, 0xff, 0xfa, 0x0f, 0xc2, 0x33, 0x90, 0xbf, 0x13, 0x68, 0xde, 0x8f, 0x4e, 0x5e, 0xd3,
-	0x8c, 0x7d, 0x08, 0x6f, 0x33, 0xe4, 0x61, 0xcc, 0x7c, 0x74, 0x47, 0xc8, 0x64, 0xd3, 0xa7, 0x93,
-	0x76, 0x45, 0xdb, 0xbf, 0x4d, 0xcc, 0x99, 0x71, 0xac, 0x67, 0xc6, 0x51, 0x4e, 0x4a, 0x8e, 0x76,
-	0xa9, 0x93, 0xf2, 0x4b, 0x15, 0x6c, 0x07, 0xa5, 0xa2, 0xdb, 0x31, 0x17, 0xe1, 0xb0, 0x8b, 0x1e,
-	0xf3, 0x1f, 0x7d, 0x26, 0x04, 0xa3, 0x0f, 0x62, 0x51, 0xb0, 0x48, 0x47, 0x70, 0x03, 0x4f, 0x29,
-	0x17, 0x34, 0xe8, 0xb9, 0xbe, 0x82, 0x71, 0xb9, 0xc2, 0x71, 0x3d, 0x0d, 0xa4, 0x2e, 0x94, 0xb4,
-	0x47, 0xde, 0xd3, 0xc1, 0x0b, 0x53, 0x4a, 0xa5, 0xc6, 0x1d, 0x68, 0x05, 0xf8, 0xe4, 0x72, 0xb0,
-	0xa4, 0xa2, 0xdb, 0x01, 0x3e, 0x59, 0x8e, 0xb3, 0xe8, 0x24, 0xea, 0x45, 0x4e, 0x62, 0x25, 0x7b,
-	0x12, 0x7d, 0x78, 0xff, 0xd2, 0x3a, 0x95, 0x7a, 0x2a, 0xcf, 0x08, 0x34, 0x0f, 0x71, 0x80, 0x2f,
-	0xdd, 0xae, 0x8b, 0xf4, 0x56, 0x8b, 0xe8, 0xad, 0x2d, 0xe8, 0xbc, 0x1c, 0x83, 0x52, 0x35, 0xfe,
-	0x48, 0xf4, 0x57, 0xe0, 0x8b, 0xd3, 0x28, 0x64, 0xa2, 0x4b, 0x83, 0x7e, 0x31, 0x91, 0x9d, 0xb9,
-	0x99, 0x4c, 0xae, 0x00, 0xf9, 0x69, 0x95, 0x29, 0xa7, 0x58, 0x85, 0xbf, 0x79, 0xf6, 0x31, 0x98,
-	0x79, 0x22, 0xa5, 0x6a, 0x4d, 0x6e, 0xb4, 0x97, 0xd5, 0xb9, 0x05, 0x0d, 0x29, 0x6b, 0x76, 0x74,
-	0x56, 0xa5, 0x41, 0x9e, 0x88, 0x7d, 0xa8, 0xee, 0xf5, 0x05, 0x8c, 0x3f, 0x82, 0xba, 0x0c, 0x4a,
-	0x79, 0x6e, 0x2c, 0xa9, 0x8e, 0xa3, 0x82, 0xec, 0x9f, 0x89, 0x3e, 0xe6, 0x32, 0xc9, 0x95, 0x78,
-	0x01, 0x1e, 0x83, 0x99, 0xe7, 0x57, 0xea, 0xd9, 0xfc, 0x46, 0xf4, 0x1d, 0xfb, 0xca, 0xfb, 0xb0,
-	0xd4, 0xb2, 0xe4, 0x39, 0x97, 0x5a, 0x16, 0x96, 0x6c, 0x3b, 0x53, 0x7c, 0x5e, 0xb8, 0x2f, 0xa6,
-	0xbb, 0x4a, 0xf5, 0xd2, 0x5d, 0xa5, 0x96, 0xdf, 0xb0, 0x36, 0x72, 0x39, 0x53, 0x51, 0x3b, 0x00,
-	0x93, 0x76, 0xd3, 0x2b, 0x56, 0x43, 0xf7, 0x5b, 0xf1, 0x7d, 0xe4, 0x71, 0x66, 0x6a, 0x5e, 0x83,
-	0xa8, 0x10, 0x9a, 0xd9, 0x94, 0xa9, 0xa6, 0x8f, 0x61, 0x45, 0x2a, 0xd0, 0xcb, 0xd7, 0xd2, 0x51,
-	0x4d, 0xa2, 0x0a, 0x6b, 0xfc, 0x93, 0x80, 0x79, 0xc7, 0xa3, 0x83, 0x70, 0x34, 0xb3, 0x7e, 0x17,
-	0xd3, 0x39, 0xdf, 0x75, 0xd5, 0xec, 0xe3, 0xe0, 0x53, 0x58, 0x17, 0x1e, 0xeb, 0xa1, 0x7c, 0x6c,
-	0xf5, 0x74, 0xf3, 0x6a, 0xe2, 0xc9, 0xbb, 0x29, 0xe9, 0x2b, 0xe9, 0xbc, 0x77, 0xe8, 0xbc, 0x99,
-	0x44, 0x27, 0xbf, 0x8d, 0x9b, 0xd0, 0xe4, 0x7d, 0x1a, 0xb9, 0x3d, 0xe6, 0xf9, 0xf8, 0x30, 0x1e,
-	0xb8, 0x0f, 0x53, 0x92, 0xaa, 0xbd, 0x57, 0x9d, 0xab, 0xd2, 0x7b, 0x37, 0x75, 0x6a, 0x01, 0xb6,
-	0x07, 0xd7, 0x17, 0x88, 0x29, 0xb3, 0xd5, 0x3f, 0xff, 0xfe, 0xec, 0xdc, 0xaa, 0x3c, 0x3f, 0xb7,
-	0x2a, 0x2f, 0xce, 0x2d, 0xf2, 0x6c, 0x6c, 0x91, 0x5f, 0xc7, 0x16, 0xf9, 0x63, 0x6c, 0x91, 0xb3,
-	0xb1, 0x45, 0xfe, 0x1a, 0x5b, 0xe4, 0xef, 0xb1, 0x55, 0x79, 0x31, 0xb6, 0xc8, 0x4f, 0x17, 0x56,
-	0xe5, 0xec, 0xc2, 0xaa, 0x3c, 0xbf, 0xb0, 0x2a, 0xdf, 0xdd, 0x14, 0xc3, 0x88, 0x0d, 0xda, 0xfe,
-	0x20, 0x8c, 0x4f, 0x3a, 0x4b, 0x1e, 0xa7, 0xb7, 0xb2, 0xb6, 0x07, 0x6f, 0xa8, 0x37, 0xdf, 0xc1,
-	0xbf, 0x01, 0x00, 0x00, 0xff, 0xff, 0x22, 0x11, 0xb6, 0x6b, 0xcf, 0x0e, 0x00, 0x00,
+	0x14, 0xf7, 0xd8, 0x0e, 0x8a, 0x5f, 0x08, 0x45, 0x4b, 0xeb, 0x6c, 0xfe, 0x2d, 0x66, 0x51, 0x51,
+	0x10, 0xc2, 0x56, 0x92, 0x5e, 0x50, 0xb9, 0x40, 0x43, 0xab, 0x4a, 0x08, 0x85, 0x35, 0xed, 0x81,
+	0xcb, 0x66, 0xbb, 0x79, 0x75, 0xa7, 0xb6, 0x77, 0xb7, 0x33, 0xb3, 0x6e, 0x54, 0x71, 0xe8, 0x8d,
+	0x2b, 0x9f, 0x80, 0x23, 0xe2, 0x8c, 0xc4, 0x77, 0x40, 0xe2, 0x92, 0x63, 0x8f, 0xc4, 0xb9, 0x70,
+	0xa3, 0x1f, 0x01, 0xcd, 0xee, 0xce, 0xda, 0xde, 0xb5, 0xc3, 0x06, 0x6d, 0x73, 0xf4, 0x7b, 0xcf,
+	0xbf, 0xf7, 0xfb, 0xbd, 0x3f, 0xb3, 0x33, 0xd0, 0x76, 0x02, 0xda, 0xf1, 0x9c, 0x21, 0xf2, 0xc0,
+	0x71, 0x91, 0x23, 0x1b, 0x51, 0x17, 0x3b, 0xa3, 0xdd, 0x0e, 0xc3, 0x67, 0x21, 0x72, 0x61, 0x33,
+	0xe4, 0x81, 0xef, 0x71, 0x6c, 0x07, 0xcc, 0x17, 0xbe, 0xb6, 0xe6, 0x04, 0xb4, 0x9d, 0x8d, 0x6f,
+	0x8f, 0x76, 0x37, 0xd6, 0x25, 0x90, 0x13, 0x8a, 0x27, 0xf2, 0xcf, 0x43, 0xe4, 0xdc, 0xe9, 0x25,
+	0xff, 0xd9, 0xd8, 0x94, 0x2e, 0xd7, 0x1f, 0x0e, 0x7d, 0x2f, 0xef, 0x34, 0x66, 0x08, 0xe4, 0xfd,
+	0x5b, 0xd2, 0x9f, 0x90, 0xc9, 0x7b, 0xa3, 0xac, 0x9c, 0x7a, 0xfd, 0x9c, 0xcb, 0xfc, 0x87, 0x40,
+	0xf3, 0x0e, 0x43, 0x47, 0xe0, 0x37, 0x0a, 0xdd, 0x8a, 0x61, 0xb4, 0x2d, 0x68, 0xa4, 0x19, 0x75,
+	0xd2, 0x22, 0x3b, 0x0d, 0x6b, 0x62, 0xd0, 0xf6, 0xa1, 0xce, 0x03, 0x74, 0xf5, 0x6a, 0x8b, 0xec,
+	0xac, 0xec, 0xbd, 0xdf, 0x9e, 0x51, 0xdc, 0x1e, 0xed, 0xb6, 0x53, 0xbc, 0x6e, 0x80, 0xae, 0x15,
+	0x05, 0x6b, 0xdb, 0x00, 0xaa, 0x62, 0xf4, 0x58, 0xaf, 0xc5, 0x98, 0x89, 0xe5, 0xfe, 0xb1, 0xe6,
+	0xc2, 0x46, 0xc8, 0x91, 0xd9, 0x29, 0x8e, 0x1d, 0x20, 0x1b, 0x52, 0xce, 0xa9, 0xef, 0x71, 0xbd,
+	0xde, 0xaa, 0xed, 0xac, 0xec, 0xdd, 0x8c, 0x32, 0xc9, 0x12, 0xca, 0x24, 0x0f, 0x38, 0xb2, 0x34,
+	0xd1, 0xe1, 0x24, 0xd8, 0xd2, 0xc3, 0x05, 0x1e, 0xd3, 0x86, 0xb5, 0x9c, 0xe0, 0xb8, 0x79, 0xda,
+	0x01, 0xbc, 0xa3, 0xe8, 0x71, 0xe1, 0x88, 0x90, 0x47, 0xb2, 0x57, 0xf6, 0xb6, 0xa3, 0x9c, 0x89,
+	0x4b, 0xa6, 0x4d, 0x4a, 0xd4, 0x8d, 0x82, 0xac, 0x55, 0x36, 0xfd, 0xd3, 0xec, 0xc2, 0x8d, 0xaf,
+	0x29, 0x17, 0x29, 0x3c, 0x57, 0x05, 0xdd, 0x84, 0x46, 0xe0, 0xf4, 0xd0, 0xe6, 0xf4, 0x45, 0x5c,
+	0xd0, 0x25, 0x6b, 0x59, 0x1a, 0xba, 0xf4, 0x05, 0xca, 0xd2, 0x44, 0x4e, 0xe1, 0xf7, 0xd1, 0x8b,
+	0xaa, 0xda, 0xb0, 0xa2, 0xf0, 0xef, 0xa4, 0xc1, 0x3c, 0x82, 0x66, 0x16, 0x34, 0x21, 0x6d, 0x00,
+	0x4c, 0x26, 0x4d, 0x27, 0xad, 0xda, 0x4e, 0xc3, 0x9a, 0xb2, 0x68, 0x1f, 0xc1, 0x35, 0x0f, 0x4f,
+	0x84, 0x9d, 0x43, 0x5f, 0x95, 0xe6, 0xc3, 0x34, 0x83, 0x05, 0xd7, 0xef, 0x61, 0xc9, 0xac, 0x7f,
+	0x80, 0x1b, 0x19, 0xcc, 0x84, 0xf4, 0xed, 0x1c, 0xe9, 0x95, 0xbd, 0xcd, 0x0b, 0x66, 0xe8, 0x7f,
+	0x29, 0xda, 0x87, 0xf7, 0xa6, 0xb3, 0x17, 0x9a, 0x6b, 0xf3, 0xdb, 0xd9, 0x32, 0xa4, 0x8c, 0x3f,
+	0xcb, 0xfe, 0xeb, 0x3f, 0x08, 0x4f, 0x41, 0xfe, 0x4e, 0xa0, 0xf9, 0x20, 0x38, 0xbe, 0xa2, 0x1d,
+	0xfb, 0x18, 0xde, 0x65, 0xc8, 0xfd, 0x90, 0xb9, 0x68, 0x8f, 0x90, 0xc9, 0xa1, 0x4f, 0x36, 0xed,
+	0x9a, 0xb2, 0x3f, 0x8c, 0xcd, 0x99, 0x75, 0xac, 0x67, 0xd6, 0x51, 0x6e, 0x4a, 0x8e, 0x76, 0xa9,
+	0x9b, 0xf2, 0x4b, 0x15, 0x4c, 0x0b, 0xa5, 0xa2, 0x3b, 0x21, 0x17, 0xfe, 0xb0, 0x8b, 0x0e, 0x73,
+	0x9f, 0x7c, 0x21, 0x04, 0xa3, 0x8f, 0x42, 0x51, 0xb0, 0x48, 0x87, 0x70, 0x13, 0x4f, 0x28, 0x17,
+	0xd4, 0xeb, 0xd9, 0x6e, 0x04, 0x63, 0xf3, 0x08, 0xc7, 0x76, 0x14, 0x50, 0x74, 0xa0, 0x24, 0x33,
+	0xf2, 0x81, 0x0a, 0x9e, 0x9b, 0x52, 0x2a, 0xd5, 0xee, 0x42, 0xcb, 0xc3, 0xe7, 0x17, 0x83, 0xc5,
+	0x15, 0xdd, 0xf2, 0xf0, 0xf9, 0x62, 0x9c, 0x79, 0x9d, 0xa8, 0x17, 0xe9, 0xc4, 0x52, 0xb6, 0x13,
+	0x7d, 0xf8, 0xf0, 0xc2, 0x3a, 0x95, 0xda, 0x95, 0x97, 0x04, 0x9a, 0x07, 0x38, 0xc0, 0x4b, 0x8f,
+	0xeb, 0x3c, 0xbd, 0xd5, 0x22, 0x7a, 0x6b, 0x73, 0x26, 0x2f, 0xc7, 0xa0, 0x54, 0x8d, 0x3f, 0x12,
+	0xf5, 0x15, 0xf8, 0xea, 0x24, 0xf0, 0x99, 0xe8, 0x52, 0xaf, 0x5f, 0x4c, 0x64, 0x67, 0x66, 0x27,
+	0xe3, 0x23, 0x40, 0x7e, 0x5a, 0x65, 0xca, 0x09, 0x56, 0xe1, 0x6f, 0x9e, 0x79, 0x04, 0x7a, 0x9e,
+	0x48, 0xa9, 0x5a, 0xe3, 0x13, 0xed, 0xb2, 0x3a, 0x37, 0xa1, 0x21, 0x65, 0x4d, 0xaf, 0xce, 0xb2,
+	0x34, 0xc8, 0x8e, 0x98, 0x07, 0xd1, 0xb9, 0x3e, 0x87, 0xf1, 0x27, 0x50, 0x97, 0x41, 0x09, 0xcf,
+	0xb5, 0x05, 0xd5, 0xb1, 0xa2, 0x20, 0xf3, 0x67, 0xa2, 0xda, 0x5c, 0x26, 0xb9, 0x12, 0x0f, 0xc0,
+	0x23, 0xd0, 0xf3, 0xfc, 0x4a, 0xed, 0xcd, 0x6f, 0x44, 0x9d, 0xb1, 0x6f, 0x7c, 0x0e, 0x4b, 0x2d,
+	0x4b, 0x9e, 0x73, 0xa9, 0x65, 0x61, 0xf1, 0x6d, 0x67, 0x82, 0xcf, 0x0b, 0xcf, 0xc5, 0xe4, 0xae,
+	0x52, 0xbd, 0xf0, 0xae, 0x52, 0xcb, 0xdf, 0xb0, 0xd6, 0x72, 0x39, 0x13, 0x51, 0xdb, 0x00, 0xe9,
+	0xb8, 0xa9, 0x2b, 0x56, 0x43, 0xcd, 0x5b, 0xf1, 0xfb, 0xc8, 0xb3, 0xcc, 0xd6, 0x5c, 0x81, 0x28,
+	0x1f, 0x9a, 0xd9, 0x94, 0x89, 0xa6, 0x4f, 0x61, 0x49, 0x2a, 0x50, 0x97, 0xaf, 0x85, 0xab, 0x1a,
+	0x47, 0x15, 0xd6, 0xf8, 0x14, 0xd6, 0x1f, 0x3a, 0x03, 0x7a, 0x15, 0x13, 0x6d, 0x6e, 0xc1, 0xc6,
+	0xbc, 0x5c, 0xb1, 0x40, 0xf3, 0x4f, 0x02, 0xfa, 0x5d, 0x87, 0x0e, 0xfc, 0xd1, 0xd4, 0x43, 0xa0,
+	0x18, 0x93, 0xd9, 0xf9, 0xaf, 0x66, 0x9f, 0x29, 0x9f, 0xc3, 0xaa, 0x70, 0x58, 0x0f, 0xe5, 0xb3,
+	0xaf, 0xa7, 0xd6, 0x48, 0x95, 0x30, 0x7e, 0xc1, 0xc5, 0x13, 0x2e, 0x9d, 0xf7, 0x0f, 0xac, 0xb7,
+	0xe3, 0xe8, 0xf8, 0xb7, 0x76, 0x0b, 0x9a, 0xbc, 0x4f, 0x03, 0xbb, 0xc7, 0x1c, 0x17, 0x1f, 0x87,
+	0x03, 0xfb, 0x71, 0x42, 0x32, 0x5a, 0xb4, 0x65, 0xeb, 0xba, 0xf4, 0xde, 0x4b, 0x9c, 0x4a, 0x80,
+	0xe9, 0xc0, 0xfa, 0x1c, 0x31, 0x65, 0x2e, 0xdd, 0x97, 0x4f, 0x4f, 0xcf, 0x8c, 0xca, 0xab, 0x33,
+	0xa3, 0xf2, 0xfa, 0xcc, 0x20, 0x2f, 0xc7, 0x06, 0xf9, 0x75, 0x6c, 0x90, 0x3f, 0xc6, 0x06, 0x39,
+	0x1d, 0x1b, 0xe4, 0xaf, 0xb1, 0x41, 0xfe, 0x1e, 0x1b, 0x95, 0xd7, 0x63, 0x83, 0xfc, 0x74, 0x6e,
+	0x54, 0x4e, 0xcf, 0x8d, 0xca, 0xab, 0x73, 0xa3, 0xf2, 0xfd, 0x2d, 0x31, 0x0c, 0xd8, 0xa0, 0xed,
+	0x0e, 0xfc, 0xf0, 0xb8, 0xb3, 0xe0, 0x99, 0x7c, 0x3b, 0x6b, 0x7b, 0xf4, 0x56, 0xf4, 0xfa, 0xdc,
+	0xff, 0x37, 0x00, 0x00, 0xff, 0xff, 0x8f, 0xff, 0xd3, 0x77, 0x59, 0x0f, 0x00, 0x00,
 }
 
 func (this *CreateNamespaceRequest) Equal(that interface{}) bool {
@@ -2411,14 +2385,14 @@ func (this *GetExportSinksResponse) Equal(that interface{}) bool {
 	}
 	return true
 }
-func (this *FailoverNamespaceRequest) Equal(that interface{}) bool {
+func (this *ValidateExportSinkRequest) Equal(that interface{}) bool {
 	if that == nil {
 		return this == nil
 	}
 
-	that1, ok := that.(*FailoverNamespaceRequest)
+	that1, ok := that.(*ValidateExportSinkRequest)
 	if !ok {
-		that2, ok := that.(FailoverNamespaceRequest)
+		that2, ok := that.(ValidateExportSinkRequest)
 		if ok {
 			that1 = &that2
 		} else {
@@ -2433,25 +2407,19 @@ func (this *FailoverNamespaceRequest) Equal(that interface{}) bool {
 	if this.Namespace != that1.Namespace {
 		return false
 	}
-	if this.RequestId != that1.RequestId {
-		return false
-	}
-	if !this.TargetRegion.Equal(that1.TargetRegion) {
-		return false
-	}
-	if this.SkipGracefulFailover != that1.SkipGracefulFailover {
+	if !this.Spec.Equal(that1.Spec) {
 		return false
 	}
 	return true
 }
-func (this *FailoverNamespaceResponse) Equal(that interface{}) bool {
+func (this *ValidateExportSinkResponse) Equal(that interface{}) bool {
 	if that == nil {
 		return this == nil
 	}
 
-	that1, ok := that.(*FailoverNamespaceResponse)
+	that1, ok := that.(*ValidateExportSinkResponse)
 	if !ok {
-		that2, ok := that.(FailoverNamespaceResponse)
+		that2, ok := that.(ValidateExportSinkResponse)
 		if ok {
 			that1 = &that2
 		} else {
@@ -2461,9 +2429,6 @@ func (this *FailoverNamespaceResponse) Equal(that interface{}) bool {
 	if that1 == nil {
 		return this == nil
 	} else if this == nil {
-		return false
-	}
-	if !this.RequestStatus.Equal(that1.RequestStatus) {
 		return false
 	}
 	return true
@@ -2791,30 +2756,25 @@ func (this *GetExportSinksResponse) GoString() string {
 	s = append(s, "}")
 	return strings.Join(s, "")
 }
-func (this *FailoverNamespaceRequest) GoString() string {
+func (this *ValidateExportSinkRequest) GoString() string {
 	if this == nil {
 		return "nil"
 	}
-	s := make([]string, 0, 8)
-	s = append(s, "&namespaceservice.FailoverNamespaceRequest{")
+	s := make([]string, 0, 6)
+	s = append(s, "&namespaceservice.ValidateExportSinkRequest{")
 	s = append(s, "Namespace: "+fmt.Sprintf("%#v", this.Namespace)+",\n")
-	s = append(s, "RequestId: "+fmt.Sprintf("%#v", this.RequestId)+",\n")
-	if this.TargetRegion != nil {
-		s = append(s, "TargetRegion: "+fmt.Sprintf("%#v", this.TargetRegion)+",\n")
+	if this.Spec != nil {
+		s = append(s, "Spec: "+fmt.Sprintf("%#v", this.Spec)+",\n")
 	}
-	s = append(s, "SkipGracefulFailover: "+fmt.Sprintf("%#v", this.SkipGracefulFailover)+",\n")
 	s = append(s, "}")
 	return strings.Join(s, "")
 }
-func (this *FailoverNamespaceResponse) GoString() string {
+func (this *ValidateExportSinkResponse) GoString() string {
 	if this == nil {
 		return "nil"
 	}
-	s := make([]string, 0, 5)
-	s = append(s, "&namespaceservice.FailoverNamespaceResponse{")
-	if this.RequestStatus != nil {
-		s = append(s, "RequestStatus: "+fmt.Sprintf("%#v", this.RequestStatus)+",\n")
-	}
+	s := make([]string, 0, 4)
+	s = append(s, "&namespaceservice.ValidateExportSinkResponse{")
 	s = append(s, "}")
 	return strings.Join(s, "")
 }
@@ -3905,7 +3865,7 @@ func (m *GetExportSinksResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) 
 	return len(dAtA) - i, nil
 }
 
-func (m *FailoverNamespaceRequest) Marshal() (dAtA []byte, err error) {
+func (m *ValidateExportSinkRequest) Marshal() (dAtA []byte, err error) {
 	size := m.Size()
 	dAtA = make([]byte, size)
 	n, err := m.MarshalToSizedBuffer(dAtA[:size])
@@ -3915,42 +3875,25 @@ func (m *FailoverNamespaceRequest) Marshal() (dAtA []byte, err error) {
 	return dAtA[:n], nil
 }
 
-func (m *FailoverNamespaceRequest) MarshalTo(dAtA []byte) (int, error) {
+func (m *ValidateExportSinkRequest) MarshalTo(dAtA []byte) (int, error) {
 	size := m.Size()
 	return m.MarshalToSizedBuffer(dAtA[:size])
 }
 
-func (m *FailoverNamespaceRequest) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+func (m *ValidateExportSinkRequest) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	i := len(dAtA)
 	_ = i
 	var l int
 	_ = l
-	if m.SkipGracefulFailover {
-		i--
-		if m.SkipGracefulFailover {
-			dAtA[i] = 1
-		} else {
-			dAtA[i] = 0
-		}
-		i--
-		dAtA[i] = 0x20
-	}
-	if m.TargetRegion != nil {
+	if m.Spec != nil {
 		{
-			size, err := m.TargetRegion.MarshalToSizedBuffer(dAtA[:i])
+			size, err := m.Spec.MarshalToSizedBuffer(dAtA[:i])
 			if err != nil {
 				return 0, err
 			}
 			i -= size
 			i = encodeVarintRequestResponse(dAtA, i, uint64(size))
 		}
-		i--
-		dAtA[i] = 0x1a
-	}
-	if len(m.RequestId) > 0 {
-		i -= len(m.RequestId)
-		copy(dAtA[i:], m.RequestId)
-		i = encodeVarintRequestResponse(dAtA, i, uint64(len(m.RequestId)))
 		i--
 		dAtA[i] = 0x12
 	}
@@ -3964,7 +3907,7 @@ func (m *FailoverNamespaceRequest) MarshalToSizedBuffer(dAtA []byte) (int, error
 	return len(dAtA) - i, nil
 }
 
-func (m *FailoverNamespaceResponse) Marshal() (dAtA []byte, err error) {
+func (m *ValidateExportSinkResponse) Marshal() (dAtA []byte, err error) {
 	size := m.Size()
 	dAtA = make([]byte, size)
 	n, err := m.MarshalToSizedBuffer(dAtA[:size])
@@ -3974,28 +3917,16 @@ func (m *FailoverNamespaceResponse) Marshal() (dAtA []byte, err error) {
 	return dAtA[:n], nil
 }
 
-func (m *FailoverNamespaceResponse) MarshalTo(dAtA []byte) (int, error) {
+func (m *ValidateExportSinkResponse) MarshalTo(dAtA []byte) (int, error) {
 	size := m.Size()
 	return m.MarshalToSizedBuffer(dAtA[:size])
 }
 
-func (m *FailoverNamespaceResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+func (m *ValidateExportSinkResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	i := len(dAtA)
 	_ = i
 	var l int
 	_ = l
-	if m.RequestStatus != nil {
-		{
-			size, err := m.RequestStatus.MarshalToSizedBuffer(dAtA[:i])
-			if err != nil {
-				return 0, err
-			}
-			i -= size
-			i = encodeVarintRequestResponse(dAtA, i, uint64(size))
-		}
-		i--
-		dAtA[i] = 0xa
-	}
 	return len(dAtA) - i, nil
 }
 
@@ -4478,7 +4409,7 @@ func (m *GetExportSinksResponse) Size() (n int) {
 	return n
 }
 
-func (m *FailoverNamespaceRequest) Size() (n int) {
+func (m *ValidateExportSinkRequest) Size() (n int) {
 	if m == nil {
 		return 0
 	}
@@ -4488,32 +4419,22 @@ func (m *FailoverNamespaceRequest) Size() (n int) {
 	if l > 0 {
 		n += 1 + l + sovRequestResponse(uint64(l))
 	}
-	l = len(m.RequestId)
-	if l > 0 {
+	if m.Spec != nil {
+		l = m.Spec.Size()
 		n += 1 + l + sovRequestResponse(uint64(l))
-	}
-	if m.TargetRegion != nil {
-		l = m.TargetRegion.Size()
-		n += 1 + l + sovRequestResponse(uint64(l))
-	}
-	if m.SkipGracefulFailover {
-		n += 2
 	}
 	return n
 }
 
-func (m *FailoverNamespaceResponse) Size() (n int) {
+func (m *ValidateExportSinkResponse) Size() (n int) {
 	if m == nil {
 		return 0
 	}
 	var l int
 	_ = l
-	if m.RequestStatus != nil {
-		l = m.RequestStatus.Size()
-		n += 1 + l + sovRequestResponse(uint64(l))
-	}
 	return n
 }
+
 
 func sovRequestResponse(x uint64) (n int) {
 	return (math_bits.Len64(x|1) + 6) / 7
@@ -4827,29 +4748,27 @@ func (this *GetExportSinksResponse) String() string {
 	}, "")
 	return s
 }
-func (this *FailoverNamespaceRequest) String() string {
+func (this *ValidateExportSinkRequest) String() string {
 	if this == nil {
 		return "nil"
 	}
-	s := strings.Join([]string{`&FailoverNamespaceRequest{`,
+	s := strings.Join([]string{`&ValidateExportSinkRequest{`,
 		`Namespace:` + fmt.Sprintf("%v", this.Namespace) + `,`,
-		`RequestId:` + fmt.Sprintf("%v", this.RequestId) + `,`,
-		`TargetRegion:` + strings.Replace(fmt.Sprintf("%v", this.TargetRegion), "RegionID", "v14.RegionID", 1) + `,`,
-		`SkipGracefulFailover:` + fmt.Sprintf("%v", this.SkipGracefulFailover) + `,`,
+		`Spec:` + strings.Replace(fmt.Sprintf("%v", this.Spec), "ExportSinkSpec", "v13.ExportSinkSpec", 1) + `,`,
 		`}`,
 	}, "")
 	return s
 }
-func (this *FailoverNamespaceResponse) String() string {
+func (this *ValidateExportSinkResponse) String() string {
 	if this == nil {
 		return "nil"
 	}
-	s := strings.Join([]string{`&FailoverNamespaceResponse{`,
-		`RequestStatus:` + strings.Replace(fmt.Sprintf("%v", this.RequestStatus), "RequestStatus", "v12.RequestStatus", 1) + `,`,
+	s := strings.Join([]string{`&ValidateExportSinkResponse{`,
 		`}`,
 	}, "")
 	return s
 }
+
 func valueToStringRequestResponse(v interface{}) string {
 	rv := reflect.ValueOf(v)
 	if rv.IsNil() {
@@ -8066,7 +7985,7 @@ func (m *GetExportSinksResponse) Unmarshal(dAtA []byte) error {
 	}
 	return nil
 }
-func (m *FailoverNamespaceRequest) Unmarshal(dAtA []byte) error {
+func (m *ValidateExportSinkRequest) Unmarshal(dAtA []byte) error {
 	l := len(dAtA)
 	iNdEx := 0
 	for iNdEx < l {
@@ -8089,10 +8008,10 @@ func (m *FailoverNamespaceRequest) Unmarshal(dAtA []byte) error {
 		fieldNum := int32(wire >> 3)
 		wireType := int(wire & 0x7)
 		if wireType == 4 {
-			return fmt.Errorf("proto: FailoverNamespaceRequest: wiretype end group for non-group")
+			return fmt.Errorf("proto: ValidateExportSinkRequest: wiretype end group for non-group")
 		}
 		if fieldNum <= 0 {
-			return fmt.Errorf("proto: FailoverNamespaceRequest: illegal tag %d (wire type %d)", fieldNum, wire)
+			return fmt.Errorf("proto: ValidateExportSinkRequest: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
 		case 1:
@@ -8129,39 +8048,7 @@ func (m *FailoverNamespaceRequest) Unmarshal(dAtA []byte) error {
 			iNdEx = postIndex
 		case 2:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field RequestId", wireType)
-			}
-			var stringLen uint64
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowRequestResponse
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				stringLen |= uint64(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			intStringLen := int(stringLen)
-			if intStringLen < 0 {
-				return ErrInvalidLengthRequestResponse
-			}
-			postIndex := iNdEx + intStringLen
-			if postIndex < 0 {
-				return ErrInvalidLengthRequestResponse
-			}
-			if postIndex > l {
-				return io.ErrUnexpectedEOF
-			}
-			m.RequestId = string(dAtA[iNdEx:postIndex])
-			iNdEx = postIndex
-		case 3:
-			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field TargetRegion", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field Spec", wireType)
 			}
 			var msglen int
 			for shift := uint(0); ; shift += 7 {
@@ -8188,33 +8075,13 @@ func (m *FailoverNamespaceRequest) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			if m.TargetRegion == nil {
-				m.TargetRegion = &v14.RegionID{}
+			if m.Spec == nil {
+				m.Spec = &v13.ExportSinkSpec{}
 			}
-			if err := m.TargetRegion.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+			if err := m.Spec.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex
-		case 4:
-			if wireType != 0 {
-				return fmt.Errorf("proto: wrong wireType = %d for field SkipGracefulFailover", wireType)
-			}
-			var v int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowRequestResponse
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				v |= int(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			m.SkipGracefulFailover = bool(v != 0)
 		default:
 			iNdEx = preIndex
 			skippy, err := skipRequestResponse(dAtA[iNdEx:])
@@ -8239,7 +8106,7 @@ func (m *FailoverNamespaceRequest) Unmarshal(dAtA []byte) error {
 	}
 	return nil
 }
-func (m *FailoverNamespaceResponse) Unmarshal(dAtA []byte) error {
+func (m *ValidateExportSinkResponse) Unmarshal(dAtA []byte) error {
 	l := len(dAtA)
 	iNdEx := 0
 	for iNdEx < l {
@@ -8262,48 +8129,12 @@ func (m *FailoverNamespaceResponse) Unmarshal(dAtA []byte) error {
 		fieldNum := int32(wire >> 3)
 		wireType := int(wire & 0x7)
 		if wireType == 4 {
-			return fmt.Errorf("proto: FailoverNamespaceResponse: wiretype end group for non-group")
+			return fmt.Errorf("proto: ValidateExportSinkResponse: wiretype end group for non-group")
 		}
 		if fieldNum <= 0 {
-			return fmt.Errorf("proto: FailoverNamespaceResponse: illegal tag %d (wire type %d)", fieldNum, wire)
+			return fmt.Errorf("proto: ValidateExportSinkResponse: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
-		case 1:
-			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field RequestStatus", wireType)
-			}
-			var msglen int
-			for shift := uint(0); ; shift += 7 {
-				if shift >= 64 {
-					return ErrIntOverflowRequestResponse
-				}
-				if iNdEx >= l {
-					return io.ErrUnexpectedEOF
-				}
-				b := dAtA[iNdEx]
-				iNdEx++
-				msglen |= int(b&0x7F) << shift
-				if b < 0x80 {
-					break
-				}
-			}
-			if msglen < 0 {
-				return ErrInvalidLengthRequestResponse
-			}
-			postIndex := iNdEx + msglen
-			if postIndex < 0 {
-				return ErrInvalidLengthRequestResponse
-			}
-			if postIndex > l {
-				return io.ErrUnexpectedEOF
-			}
-			if m.RequestStatus == nil {
-				m.RequestStatus = &v12.RequestStatus{}
-			}
-			if err := m.RequestStatus.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
-				return err
-			}
-			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
 			skippy, err := skipRequestResponse(dAtA[iNdEx:])
@@ -8328,6 +8159,7 @@ func (m *FailoverNamespaceResponse) Unmarshal(dAtA []byte) error {
 	}
 	return nil
 }
+
 func skipRequestResponse(dAtA []byte) (n int, err error) {
 	l := len(dAtA)
 	iNdEx := 0
