@@ -29,37 +29,38 @@ func init() {
 }
 
 var fileDescriptor_d746e5fd89aff5eb = []byte{
-	// 471 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x9c, 0x95, 0xbf, 0xae, 0xd3, 0x30,
-	0x14, 0x87, 0xe3, 0x85, 0xc1, 0xe2, 0x72, 0x2f, 0x5e, 0x90, 0x2e, 0x92, 0x07, 0x24, 0x36, 0xae,
-	0x43, 0x2f, 0x6c, 0x97, 0x05, 0xca, 0x9f, 0x05, 0x31, 0xb4, 0x62, 0x61, 0x41, 0x69, 0x7a, 0x24,
-	0x02, 0x49, 0x63, 0x6c, 0x27, 0x02, 0x89, 0x81, 0x47, 0x60, 0xe6, 0x09, 0x78, 0x07, 0x5e, 0x80,
-	0xb1, 0x63, 0x47, 0x9a, 0x2e, 0x8c, 0x7d, 0x04, 0x54, 0x52, 0xbb, 0x8e, 0xa3, 0x24, 0x2e, 0x5b,
-	0x55, 0x7f, 0x3f, 0x7f, 0xf1, 0xf1, 0xc9, 0x09, 0xbe, 0x1b, 0xf1, 0x24, 0x5c, 0x44, 0x19, 0x48,
-	0x1e, 0xc5, 0x20, 0x41, 0x94, 0x49, 0x0c, 0x61, 0x39, 0x0a, 0xf7, 0x3f, 0x19, 0x17, 0xb9, 0xca,
-	0xc9, 0xad, 0x88, 0x27, 0xcc, 0xc5, 0x58, 0x39, 0x3a, 0x67, 0x5d, 0x79, 0x01, 0x1f, 0x0b, 0x90,
-	0xea, 0xad, 0x00, 0xc9, 0xf3, 0x85, 0xdc, 0x6f, 0x74, 0xf9, 0xf3, 0x04, 0x9f, 0xbd, 0xd2, 0xf8,
-	0xb4, 0xc6, 0x49, 0x89, 0x4f, 0xc7, 0x02, 0x22, 0x05, 0x66, 0x85, 0x84, 0xac, 0xc3, 0xc8, 0x1c,
-	0x72, 0x52, 0x7b, 0xce, 0xef, 0xfb, 0x07, 0xea, 0x07, 0xba, 0x13, 0x10, 0x89, 0x6f, 0xbc, 0x4c,
-	0xa4, 0x32, 0x4b, 0x92, 0xb0, 0xce, 0x5d, 0x9a, 0xa0, 0xb6, 0x86, 0xde, 0xbc, 0x91, 0x72, 0x7c,
-	0xf2, 0x02, 0x6c, 0xe7, 0x45, 0xe7, 0x1e, 0x0d, 0x4e, 0x2b, 0x99, 0x2f, 0x6e, 0x8c, 0x19, 0xbe,
-	0x6e, 0x2f, 0x91, 0x7b, 0x5e, 0x3b, 0x68, 0xdf, 0x85, 0x27, 0x6d, 0x74, 0x25, 0x3e, 0x7d, 0xcd,
-	0xe7, 0x9e, 0xb7, 0xe9, 0x90, 0xc3, 0xb7, 0xd9, 0x0a, 0x18, 0xef, 0x77, 0x84, 0x6f, 0x4f, 0x60,
-	0x97, 0x19, 0x17, 0x52, 0xe5, 0xd9, 0x14, 0x22, 0x11, 0xbf, 0x7b, 0xac, 0x94, 0x48, 0x66, 0x85,
-	0x02, 0x72, 0xd5, 0xb9, 0x67, 0x4f, 0x4a, 0x3f, 0xd0, 0xa3, 0xff, 0x0b, 0xdb, 0x45, 0x79, 0x0a,
-	0x29, 0xf8, 0x15, 0xc5, 0x21, 0x87, 0x8b, 0xd2, 0x0a, 0x18, 0xef, 0x67, 0x7c, 0x56, 0xf7, 0xff,
-	0xb3, 0x4f, 0x3c, 0x17, 0x6a, 0x9a, 0x2c, 0x3e, 0x90, 0xa1, 0x57, 0xe5, 0x80, 0x6a, 0xf3, 0xe8,
-	0x88, 0x84, 0xd3, 0xe8, 0x96, 0xb7, 0xb7, 0x93, 0xda, 0x52, 0xe6, 0x8b, 0xdb, 0x87, 0xad, 0x2b,
-	0xe1, 0x75, 0x58, 0x17, 0x1d, 0x3e, 0x6c, 0x3b, 0x61, 0xab, 0xeb, 0xce, 0xf4, 0x52, 0xbb, 0xe8,
-	0xb0, 0xba, 0x9d, 0xb0, 0x5b, 0x6b, 0x37, 0x6c, 0x0e, 0x6b, 0x92, 0xf4, 0x8f, 0x25, 0x8b, 0x1c,
-	0x6e, 0xad, 0x56, 0xc0, 0x9e, 0x9e, 0x8d, 0x8b, 0xe8, 0x9b, 0x9e, 0x4d, 0x70, 0x78, 0x7a, 0xba,
-	0xbc, 0x91, 0x7e, 0xc1, 0x37, 0x9f, 0x47, 0x49, 0x9a, 0x97, 0x20, 0x0e, 0x6f, 0x52, 0x77, 0xd9,
-	0x5a, 0xac, 0x56, 0x5f, 0x1e, 0x13, 0xd1, 0xf6, 0x27, 0xef, 0x97, 0x6b, 0x1a, 0xac, 0xd6, 0x34,
-	0xd8, 0xae, 0x29, 0xfa, 0x5a, 0x51, 0xf4, 0xa3, 0xa2, 0xe8, 0x57, 0x45, 0xd1, 0xb2, 0xa2, 0xe8,
-	0x77, 0x45, 0xd1, 0x9f, 0x8a, 0x06, 0xdb, 0x8a, 0xa2, 0x6f, 0x1b, 0x1a, 0x2c, 0x37, 0x34, 0x58,
-	0x6d, 0x68, 0xf0, 0xe6, 0xa1, 0xca, 0xb8, 0x48, 0x59, 0x9c, 0xe6, 0xc5, 0x3c, 0xec, 0xf8, 0x5e,
-	0x5e, 0xb9, 0xff, 0xcd, 0xae, 0xfd, 0xfb, 0x60, 0x3e, 0xf8, 0x1b, 0x00, 0x00, 0xff, 0xff, 0x3f,
-	0x9f, 0x8e, 0x77, 0xa2, 0x07, 0x00, 0x00,
+	// 492 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x9c, 0x95, 0x3f, 0x6f, 0x13, 0x31,
+	0x18, 0x87, 0xef, 0x16, 0x06, 0x0b, 0xda, 0xe2, 0x05, 0xa9, 0x48, 0x1e, 0x90, 0xd8, 0xa8, 0x8f,
+	0xa4, 0x6c, 0x65, 0x81, 0xf2, 0x67, 0x41, 0x0c, 0x8d, 0x60, 0x60, 0x41, 0x6e, 0xfa, 0x4a, 0x18,
+	0x2e, 0x39, 0x63, 0x3b, 0x27, 0x90, 0x90, 0xe0, 0x23, 0x30, 0xf3, 0x09, 0xf8, 0x10, 0x7c, 0x00,
+	0xc6, 0x8c, 0x1d, 0xc9, 0x65, 0x61, 0xec, 0x47, 0x40, 0xe5, 0x62, 0xc7, 0xb1, 0xe5, 0x3b, 0xa7,
+	0x5b, 0x14, 0x3f, 0xbf, 0xf7, 0x39, 0xdb, 0xef, 0xbd, 0x87, 0xee, 0x32, 0xc1, 0x8b, 0x29, 0x9b,
+	0x80, 0x12, 0x6c, 0x0c, 0x0a, 0x64, 0xcd, 0xc7, 0x50, 0xd4, 0x83, 0x62, 0xf5, 0x93, 0x0a, 0x59,
+	0xe9, 0x0a, 0xdf, 0x62, 0x82, 0x53, 0x1f, 0xa3, 0xf5, 0x60, 0x9f, 0xc6, 0xf2, 0x12, 0x3e, 0xce,
+	0x40, 0xe9, 0xb7, 0x12, 0x94, 0xa8, 0xa6, 0x6a, 0x55, 0x68, 0xf8, 0x6b, 0x07, 0xed, 0xbd, 0x34,
+	0xf8, 0xa8, 0xc5, 0x71, 0x8d, 0x76, 0x8f, 0x25, 0x30, 0x0d, 0x76, 0x05, 0x17, 0x34, 0x62, 0xa4,
+	0x1e, 0x79, 0xd2, 0x7a, 0xf6, 0xef, 0xa7, 0x07, 0xda, 0x07, 0xba, 0x93, 0x61, 0x85, 0x76, 0x5e,
+	0x70, 0xa5, 0xed, 0x92, 0xc2, 0x34, 0x5a, 0x65, 0x13, 0x34, 0xd6, 0x22, 0x99, 0xb7, 0x52, 0x81,
+	0x6e, 0x3c, 0x07, 0xd7, 0x79, 0x10, 0xad, 0xb1, 0xc1, 0x19, 0x25, 0x4d, 0xc5, 0xad, 0x71, 0x82,
+	0xae, 0xbb, 0x4b, 0xf8, 0x5e, 0x52, 0x05, 0xe3, 0x3b, 0x48, 0xa4, 0xad, 0xae, 0x46, 0xbb, 0xaf,
+	0xc4, 0x59, 0xe2, 0x6d, 0x7a, 0x64, 0xff, 0x6d, 0x06, 0x01, 0xeb, 0xfd, 0x91, 0xa3, 0xdb, 0x27,
+	0x70, 0x99, 0x39, 0x9e, 0x29, 0x5d, 0x4d, 0x46, 0xc0, 0xe4, 0xf8, 0xdd, 0x23, 0xad, 0x25, 0x3f,
+	0x9d, 0x69, 0xc0, 0x47, 0xd1, 0x9a, 0x1d, 0x29, 0xf3, 0x40, 0x0f, 0xaf, 0x16, 0x76, 0x0f, 0xe5,
+	0x09, 0x94, 0x90, 0x76, 0x28, 0x1e, 0xd9, 0x7f, 0x28, 0x41, 0xc0, 0x7a, 0x3f, 0xa3, 0xbd, 0xb6,
+	0xff, 0x9f, 0x7e, 0x12, 0x95, 0xd4, 0x23, 0x3e, 0xfd, 0x80, 0xfb, 0x5e, 0x95, 0x35, 0x6a, 0xcc,
+	0x83, 0x2d, 0x12, 0x5e, 0xa3, 0x3b, 0xde, 0xce, 0x4e, 0x0a, 0xa5, 0x34, 0x15, 0x77, 0x37, 0xdb,
+	0x9e, 0x44, 0xd2, 0x66, 0x7d, 0xb4, 0x7f, 0xb3, 0x61, 0xc2, 0x55, 0xb7, 0x9d, 0x99, 0xa4, 0xf6,
+	0xd1, 0x7e, 0x75, 0x98, 0x70, 0x5b, 0xeb, 0x72, 0xd8, 0xac, 0xd7, 0x14, 0xee, 0x1e, 0x4b, 0x0e,
+	0xd9, 0xdf, 0x5a, 0x41, 0xc0, 0x7a, 0xbf, 0x22, 0xfc, 0x9a, 0x95, 0xdc, 0xdb, 0xf4, 0x30, 0x5a,
+	0x29, 0x84, 0x8d, 0xfd, 0x70, 0xab, 0x8c, 0x3b, 0xbe, 0x37, 0x3a, 0xa1, 0x6b, 0x7c, 0x6f, 0x82,
+	0xfd, 0xe3, 0xdb, 0xe7, 0xad, 0xf4, 0x0b, 0xba, 0xf9, 0x8c, 0xf1, 0xb2, 0xaa, 0x41, 0xae, 0x5f,
+	0xe5, 0xf8, 0xbd, 0x05, 0xac, 0x51, 0x0f, 0xb7, 0x89, 0x18, 0xfb, 0xe3, 0xf7, 0xf3, 0x05, 0xc9,
+	0xce, 0x17, 0x24, 0xbb, 0x58, 0x90, 0xfc, 0x5b, 0x43, 0xf2, 0x9f, 0x0d, 0xc9, 0x7f, 0x37, 0x24,
+	0x9f, 0x37, 0x24, 0xff, 0xd3, 0x90, 0xfc, 0x6f, 0x43, 0xb2, 0x8b, 0x86, 0xe4, 0xdf, 0x97, 0x24,
+	0x9b, 0x2f, 0x49, 0x76, 0xbe, 0x24, 0xd9, 0x9b, 0x07, 0x7a, 0x22, 0x64, 0x49, 0xc7, 0x65, 0x35,
+	0x3b, 0x2b, 0x22, 0x1f, 0xec, 0x23, 0xff, 0xbf, 0xd3, 0x6b, 0xff, 0xbf, 0xd8, 0x87, 0xff, 0x02,
+	0x00, 0x00, 0xff, 0xff, 0xde, 0x54, 0xaa, 0xf1, 0x23, 0x08, 0x00, 0x00,
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -98,10 +99,10 @@ type NamespaceServiceClient interface {
 	UpdateExportSink(ctx context.Context, in *UpdateExportSinkRequest, opts ...grpc.CallOption) (*UpdateExportSinkResponse, error)
 	// ListExportSinks lists the export sinks under the specified namespace on Temporal Cloud
 	ListExportSinks(ctx context.Context, in *ListExportSinksRequest, opts ...grpc.CallOption) (*ListExportSinksResponse, error)
+	// ValidateExportSink that could write test file to sink on Temporal Cloud
+	ValidateExportSink(ctx context.Context, in *ValidateExportSinkRequest, opts ...grpc.CallOption) (*ValidateExportSinkResponse, error)
 	// GetExportSinks retrieves the export sinks under the specified namespace on Temporal Cloud
 	GetExportSinks(ctx context.Context, in *GetExportSinksRequest, opts ...grpc.CallOption) (*GetExportSinksResponse, error)
-	// FailoverNamespace failovers the namespace from the source_region to the target_region on Temporal Cloud
-	FailoverNamespace(ctx context.Context, in *FailoverNamespaceRequest, opts ...grpc.CallOption) (*FailoverNamespaceResponse, error)
 }
 
 type namespaceServiceClient struct {
@@ -220,18 +221,18 @@ func (c *namespaceServiceClient) ListExportSinks(ctx context.Context, in *ListEx
 	return out, nil
 }
 
-func (c *namespaceServiceClient) GetExportSinks(ctx context.Context, in *GetExportSinksRequest, opts ...grpc.CallOption) (*GetExportSinksResponse, error) {
-	out := new(GetExportSinksResponse)
-	err := c.cc.Invoke(ctx, "/api.namespaceservice.v1.NamespaceService/GetExportSinks", in, out, opts...)
+func (c *namespaceServiceClient) ValidateExportSink(ctx context.Context, in *ValidateExportSinkRequest, opts ...grpc.CallOption) (*ValidateExportSinkResponse, error) {
+	out := new(ValidateExportSinkResponse)
+	err := c.cc.Invoke(ctx, "/api.namespaceservice.v1.NamespaceService/ValidateExportSink", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *namespaceServiceClient) FailoverNamespace(ctx context.Context, in *FailoverNamespaceRequest, opts ...grpc.CallOption) (*FailoverNamespaceResponse, error) {
-	out := new(FailoverNamespaceResponse)
-	err := c.cc.Invoke(ctx, "/api.namespaceservice.v1.NamespaceService/FailoverNamespace", in, out, opts...)
+func (c *namespaceServiceClient) GetExportSinks(ctx context.Context, in *GetExportSinksRequest, opts ...grpc.CallOption) (*GetExportSinksResponse, error) {
+	out := new(GetExportSinksResponse)
+	err := c.cc.Invoke(ctx, "/api.namespaceservice.v1.NamespaceService/GetExportSinks", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -264,10 +265,10 @@ type NamespaceServiceServer interface {
 	UpdateExportSink(context.Context, *UpdateExportSinkRequest) (*UpdateExportSinkResponse, error)
 	// ListExportSinks lists the export sinks under the specified namespace on Temporal Cloud
 	ListExportSinks(context.Context, *ListExportSinksRequest) (*ListExportSinksResponse, error)
+	// ValidateExportSink that could write test file to sink on Temporal Cloud
+	ValidateExportSink(context.Context, *ValidateExportSinkRequest) (*ValidateExportSinkResponse, error)
 	// GetExportSinks retrieves the export sinks under the specified namespace on Temporal Cloud
 	GetExportSinks(context.Context, *GetExportSinksRequest) (*GetExportSinksResponse, error)
-	// FailoverNamespace failovers the namespace from the source_region to the target_region on Temporal Cloud
-	FailoverNamespace(context.Context, *FailoverNamespaceRequest) (*FailoverNamespaceResponse, error)
 }
 
 // UnimplementedNamespaceServiceServer can be embedded to have forward compatible implementations.
@@ -310,11 +311,11 @@ func (*UnimplementedNamespaceServiceServer) UpdateExportSink(ctx context.Context
 func (*UnimplementedNamespaceServiceServer) ListExportSinks(ctx context.Context, req *ListExportSinksRequest) (*ListExportSinksResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListExportSinks not implemented")
 }
+func (*UnimplementedNamespaceServiceServer) ValidateExportSink(ctx context.Context, req *ValidateExportSinkRequest) (*ValidateExportSinkResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method ValidateExportSink not implemented")
+}
 func (*UnimplementedNamespaceServiceServer) GetExportSinks(ctx context.Context, req *GetExportSinksRequest) (*GetExportSinksResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetExportSinks not implemented")
-}
-func (*UnimplementedNamespaceServiceServer) FailoverNamespace(ctx context.Context, req *FailoverNamespaceRequest) (*FailoverNamespaceResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method FailoverNamespace not implemented")
 }
 
 func RegisterNamespaceServiceServer(s *grpc.Server, srv NamespaceServiceServer) {
@@ -537,6 +538,24 @@ func _NamespaceService_ListExportSinks_Handler(srv interface{}, ctx context.Cont
 	return interceptor(ctx, in, info, handler)
 }
 
+func _NamespaceService_ValidateExportSink_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(ValidateExportSinkRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(NamespaceServiceServer).ValidateExportSink(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/api.namespaceservice.v1.NamespaceService/ValidateExportSink",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(NamespaceServiceServer).ValidateExportSink(ctx, req.(*ValidateExportSinkRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 func _NamespaceService_GetExportSinks_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(GetExportSinksRequest)
 	if err := dec(in); err != nil {
@@ -551,24 +570,6 @@ func _NamespaceService_GetExportSinks_Handler(srv interface{}, ctx context.Conte
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(NamespaceServiceServer).GetExportSinks(ctx, req.(*GetExportSinksRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _NamespaceService_FailoverNamespace_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(FailoverNamespaceRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(NamespaceServiceServer).FailoverNamespace(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/api.namespaceservice.v1.NamespaceService/FailoverNamespace",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(NamespaceServiceServer).FailoverNamespace(ctx, req.(*FailoverNamespaceRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -626,12 +627,12 @@ var _NamespaceService_serviceDesc = grpc.ServiceDesc{
 			Handler:    _NamespaceService_ListExportSinks_Handler,
 		},
 		{
-			MethodName: "GetExportSinks",
-			Handler:    _NamespaceService_GetExportSinks_Handler,
+			MethodName: "ValidateExportSink",
+			Handler:    _NamespaceService_ValidateExportSink_Handler,
 		},
 		{
-			MethodName: "FailoverNamespace",
-			Handler:    _NamespaceService_FailoverNamespace_Handler,
+			MethodName: "GetExportSinks",
+			Handler:    _NamespaceService_GetExportSinks_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},
